Add 60+ GTNH ores: short-ID migration, 79 ore blocks, 60 small ores, textures, veins, drops, map colors
- BlockType: byte→short ID migration (BY_ID expands to 4096 slots)
- Add 59 new ore minerals (tiles 120-179) with CRUSHED_ items
- Add 40 new SMALL_*_ORE indicators (tiles 180-219)
- Chunk/ChunkStorage: byte[]→short[] blocks/overlays with new magic-byte
  format (0x53) and backward-compat read of old byte-format saves
- TextureAtlas: 100 new paintOreTile calls for ore and small-ore tiles
- GthnOreGenerator: new vein and small-ore spawn definitions
- MapRenderer: reg() calls for all new ore minerals (map legend colors)
- OreDrops: drop registrations for all new ore and small-ore blocks
- All 375 tests pass

+++ src/main/java/com/minecraftclone/world/Furnace.java
@@ -261,7 +261,7 @@ public class Furnace implements BlockEntity, StorageContainer {
                 types[i] = null;
                 counts[i] = 0;
             } else {
-                types[i] = BlockType.byId((byte) id);
+                types[i] = BlockType.byId(id);
                 counts[i] = count;
             }
         }
